* commit 'fec3a3971baed63f05b59f0641d9291546a1fd63':
  drivers/crypto: enable esn for dpaa/dpaa2/caam_jr
  test/cryptodev: improve pdcp test presentation
  ring: enforce reading tail before slots
  PVT: examples/l3fwd: fix the split mode option dpdmux connection naming
  config: increase maximum number of raw devices to 64
  PVT: drivers: fix 32 bit and various compilation issues
  net/dpaa2: fix Max Rx length as per HW MC settings
  PVT: net/dpaa2: pre-allocate aligned fqids for rx queues
  net/dpaa2: use lfqid at additional places in tx path
  PVT: net/dpaa2: make parse error drops to be configurable

+++ drivers/mempool/dpaa2/dpaa2_hw_mempool.c
@@ -414,8 +414,8 @@ dpaa2_populate(struct rte_mempool *mp, unsigned int max_objs,
 
 	if (!msl) {
 		DPAA2_MEMPOOL_DEBUG("Memsegment is External.\n");
-		rte_fslmc_vfio_mem_dmamap((uint64_t)vaddr,
-				(uint64_t)paddr, (uint64_t)len);
+		rte_fslmc_vfio_mem_dmamap((size_t)vaddr,
+				(size_t)paddr, (size_t)len);
 	}
 	/* Insert entry into the PA->VA Table */
 	dpaax_iova_table_update(paddr, vaddr, len);
